图形调试器为所有可见对象具名化
图形调试器为所有可见对象具名化

+++ 23 Dynamic Cube Mapping/Model.cpp
@@ -148,7 +148,7 @@ void Model::SetDebugObjectName(const std::string& name)
 #if (defined(DEBUG) || defined(_DEBUG)) && (GRAPHICS_DEBUGGER_OBJECT_NAME)
 
 	size_t modelPartSize = modelParts.size();
-	for (int i = 0; i < modelPartSize; ++i)
+	for (size_t i = 0; i < modelPartSize; ++i)
 	{
 		std::string vbName = name + ".part[" + std::to_string(i) + "].VertexBuffer";
 		std::string ibName = name + ".part[" + std::to_string(i) + "].IndexBuffer";
